backend/requirements.txt
# Core FastAPI Dependencies
fastapi==0.109.0
uvicorn[standard]==0.27.0
motor==3.3.2
pymongo==4.6.1
pydantic==2.5.3
pydantic[email]==2.5.3
python-multipart==0.0.6

# Authentication Dependencies
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# OCR Dependencies
paddleocr==2.7.0.3
paddlepaddle==2.5.2
opencv-python==4.8.1.78
Pillow==10.1.0

# NLP Dependencies
spacy==3.7.2
dateparser==1.2.0

# Additional utilities
numpy==1.24.3
requests==2.31.0